chore: complete split of wasm and non-wasm
