+++ core/service/src/engine/threshold/mod.rs
@@ -1,6 +1,4 @@
 mod endpoint;
 pub mod service;
 pub mod threshold_kms;
-#[cfg(any(test, feature = "testing"))]
-pub mod threshold_kms_mock;
 pub mod traits;
